+++ docs/CHANGELOG.md
@@ -68,6 +68,7 @@
     * `Fix Eggs Spawning Abnormally Healthy Chickens` (Enabled By Default)
     * `Correctly Close API Sockets` (Enabled By Default)
     * `Optimized API Sockets` (Enabled By Default)
+    * `Stop Doors And Trapdoors Closing When Updated` (Disabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`